+++ src/components/Text/Text.test.tsx
@@ -48,7 +48,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -78,7 +77,6 @@ describe('Text', () => {
         timeZone: '',
         replaceVariables,
         eventBus: eventBus as any,
-        htmlContents: [],
       };
 
       await act(async () => render(<Text {...props} />));
@@ -113,7 +111,6 @@ describe('Text', () => {
         timeZone: '',
         replaceVariables,
         eventBus: eventBus as any,
-        htmlContents: [],
       };
 
       await act(async () => render(<Text {...props} />));
@@ -149,7 +146,6 @@ describe('Text', () => {
         timeZone: '',
         replaceVariables,
         eventBus: eventBus as any,
-        htmlContents: [],
       };
 
       const { rerender } = await act(async () => render(<Text {...props} />));
@@ -191,7 +187,6 @@ describe('Text', () => {
         timeZone: '',
         replaceVariables,
         eventBus: eventBus as any,
-        htmlContents: [],
       };
 
       await act(async () => render(<Text {...props} />));
@@ -235,7 +230,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -284,7 +278,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -344,7 +337,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -383,7 +375,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -425,7 +416,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -456,7 +446,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -496,7 +485,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -547,7 +535,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));
@@ -610,7 +597,6 @@ describe('Text', () => {
       timeZone: '',
       replaceVariables: (str: string) => str,
       eventBus: {} as any,
-      htmlContents: [],
     };
 
     await act(async () => render(<Text {...props} />));