+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useContentPartials';
 export * from './useExternalResources';