+++ src/utils/index.ts
@@ -3,5 +3,6 @@ export * from './code-parameters';
 export * from './external-resources';
 export * from './handlebars';
 export * from './html';
+export * from './partials';
 export * from './rows';
 export * from './variable';